chore: simplify recorder classes
